requirements.txt
openai
python-dotenv
pypdf2
python-docx